CIP-2067 - update telmodel to v1.11.2 to change epoch from int to float

+++ src/ska_mid_cbf_mcs/subarray/subarray_component_manager.py
@@ -660,7 +660,7 @@ class CbfSubarrayComponentManager(
                     t = Thread(
                         target=self._update_timing_beam_weights,
                         args=(
-                            int(timing_beam_weights["epoch"]),
+                            float(timing_beam_weights["epoch"]),
                             json.dumps(
                                 timing_beam_weights[
                                     "timing_beam_weights_details"
